Panning example changes
Replaced the coin animation images with a more appropriate artwork that is more likely to be reusable in possible future examples.
*The artwork is created by myself and I am purposefully submitting it.

Made some changes to the text in details area of pan.script to fit better in webpage and make it easier to read. (without having to scroll so much back and forth)